requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.4.4; python_version >= '3.8'
aiohttp==3.11.11; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
attrs==25.1.0; python_version >= '3.8'
bitarray==3.0.0
certifi==2024.12.14; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
ckzg==2.0.1
cytoolz==1.0.1; python_version >= '3.8'
distro==1.9.0; python_version >= '3.6'
eth-abi==5.2.0; python_version >= '3.8' and python_version < '4'
eth-account==0.13.4; python_version >= '3.8' and python_version < '4'
eth-hash[pycryptodome]==0.7.1; python_version >= '3.8' and python_version < '4'
eth-keyfile==0.8.1; python_version >= '3.8' and python_version < '4'
eth-keys==0.6.1; python_version >= '3.8' and python_version < '4'
eth-rlp==2.1.0; python_version >= '3.8' and python_version < '4'
eth-typing==5.1.0; python_version >= '3.8' and python_version < '4'
eth-utils==5.2.0; python_version >= '3.8' and python_version < '4'
frozenlist==1.5.0; python_version >= '3.8'
greenlet==3.1.1; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
hexbytes==1.3.0; python_version >= '3.8' and python_version < '4'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jiter==0.8.2; python_version >= '3.8'
multidict==6.1.0; python_version >= '3.8'
openai==1.60.2; python_version >= '3.8'
parsimonious==0.10.0
pillow==11.1.0; python_version >= '3.9'
propcache==0.2.1; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
pycryptodome==3.21.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pysocks==1.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pyunormalize==16.0.0; python_version >= '3.6'
pyyaml==6.0.2; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rlp==4.0.1; python_version >= '3.8' and python_version < '4'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.37; python_version >= '3.7'
toolz==1.0.0; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
types-requests==2.32.0.20241016; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.9'
web3==7.7.0; python_version >= '3.8' and python_version < '4'
websockets==13.1; python_version >= '3.8'
yarl==1.18.3; python_version >= '3.9'
